feat: update runscript & finalization
